change parallaxscrollview to themedscrollview
