+++ components/ThemedScrollView.tsx
@@ -1,0 +1,38 @@
+import type { PropsWithChildren } from "react";
+import { StyleSheet } from "react-native";
+import Animated, { useAnimatedRef } from "react-native-reanimated";
+
+import { ThemedView } from "@/components/ThemedView";
+import { useBottomTabOverflow } from "@/components/ui/TabBarBackground";
+
+type Props = PropsWithChildren<{}>;
+
+export default function ThemedScrollView({ children }: Props) {
+    const scrollRef = useAnimatedRef<Animated.ScrollView>();
+    const bottom = useBottomTabOverflow();
+
+    return (
+        <ThemedView style={styles.container}>
+            <Animated.ScrollView
+                ref={scrollRef}
+                scrollEventThrottle={16}
+                scrollIndicatorInsets={{ bottom }}
+                contentContainerStyle={{ paddingBottom: bottom }}
+            >
+                <ThemedView style={styles.content}>{children}</ThemedView>
+            </Animated.ScrollView>
+        </ThemedView>
+    );
+}
+
+const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+    },
+    content: {
+        flex: 1,
+        padding: 32,
+        gap: 16,
+        overflow: "hidden",
+    },
+});
